# openapi.yaml
openapi: 3.0.3
info:
  title: AZAAM International Medics Network API
  version: 1.0.0
  description: Versioned REST contract for the clinical attachment platform. Organization-specific legal and accreditation claims are TO BE CONFIRMED.
servers:
  - url: https://api.example.com/api/v1
    description: Production domain placeholder
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /applications/{id}/submit:
    post:
      tags: [Applications]
      summary: Submit a complete application draft
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /applications/{id}/approve:
    post:
      tags: [Applications]
      summary: Approve an application
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '409': { $ref: '#/components/responses/ConflictError' }
  /applications/{id}/reject:
    post:
      tags: [Applications]
      summary: Reject an application with a required reason
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id: { in: path, name: id, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    Error:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        errors: { type: array, items: { type: object } }
  responses:
    Success:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            required: [success, data]
            properties:
              success: { type: boolean, example: true }
              data: {}
              message: { type: string }
    Error:
      description: Error response
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidationError: { $ref: '#/components/responses/Error' }
...
    ForbiddenError: { $ref: '#/components/responses/Error' }
    ConflictError: { $ref: '#/components/responses/Error' }
    NotFoundError: { $ref: '#/components/responses/Error' }